JP-1122 Revise persistence Device... First batch of changes (need to commit for laptop switch)

+++ comu/com.energyict.mdc.protocol.api/src/main/java/com/energyict/mdc/protocol/api/device/offline/OfflineLoadProfile.java
@@ -55,7 +55,7 @@ public interface OfflineLoadProfile extends Offline {
      *
      * @return the ID of the Device.
      */
-    public int getDeviceId ();
+    public long getDeviceId ();
 
     /**
      * Returns the SerialNumber of the Master Device
